feat: added UserGameSession repository and declared UserGameSession interface and controller

+++ gamification-service/GamificationService/src/main/java/com/carlos/gamificationservice/repository/UserStreakRepository.java
@@ -3,9 +3,7 @@ package com.carlos.gamificationservice.repository;
 import com.carlos.gamificationservice.models.UserStreak;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserStreakRepository extends JpaRepository<UserStreak, Integer> {
+public interface UserStreakRepository extends JpaRepository<UserStreak, Long> {
     UserStreak getUserStreakByUserName(String userName);
     // Here we must declare the methods for the db interaction.
-
-
 }
